Move credentials stuff to the credentials package
