+++ src/main/resources/index.jelly
@@ -24,5 +24,5 @@ THE SOFTWARE.
 -->
 <?jelly escape-by-default='true'?>
 <div>
-  Provides a common shared functionality for various Docker-related plugins.
+  Provides the common shared functionality for various Docker-related plugins.
 </div>
