+++ src/main/resources/org/jenkinsci/plugins/docker/commons/credentials/DockerRegistryEndpoint/config.jelly
