+++ src/main/resources/org/jenkinsci/plugins/docker/commons/credentials/DockerRegistryEndpoint/help-url.html
