+++ src/main/resources/org/jenkinsci/plugins/docker/commons/credentials/DockerServerCredentials/credentials.jelly
