+++ src/main/resources/org/jenkinsci/plugins/docker/commons/credentials/DockerServerDomainSpecification/help.jelly
