+++ src/main/resources/org/jenkinsci/plugins/docker/commons/credentials/DockerServerEndpoint/config.jelly
