+++ src/test/java/org/jenkinsci/plugins/docker/commons/ConfigTest.java
@@ -24,6 +24,9 @@
 
 package org.jenkinsci.plugins.docker.commons;
 
+import org.jenkinsci.plugins.docker.commons.credentials.DockerServerCredentials;
+import org.jenkinsci.plugins.docker.commons.credentials.DockerServerEndpoint;
+import org.jenkinsci.plugins.docker.commons.credentials.DockerRegistryEndpoint;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
 import com.cloudbees.plugins.credentials.CredentialsScope;
 import com.cloudbees.plugins.credentials.CredentialsStore;
@@ -35,6 +38,7 @@ import java.util.Collections;
 import org.junit.Test;
 import org.junit.Rule;
 import org.jvnet.hudson.test.JenkinsRule;
+
 
 public class ConfigTest {
